pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "openai>=2.31.0",
    "openpyxl>=3.1.5",
    "pydantic-ai>=1.82.0",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.26",
    "sqlalchemy[asyncio]>=2.0.49",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.44.0",
]
